+++ tests/testthat/test-list_helpers.R
@@ -2,7 +2,7 @@ test_that("initialization of Index works", {
   checkmate::expect_r6(Index$new(), "Index")
 })
 
-test_that("basic example works", {
+test_that("basic Index example works", {
   my_index <- Index$new()
   my_index$
     add(42, c("number", "rational"))$
@@ -29,4 +29,21 @@ test_that("basic example works", {
   my_index$remove(ids = 4, shift_ids = FALSE)
 })
 
-
+test_that("lists can be merged", {
+  expect_equal(
+    merge_lists(list("a" = 1, "b" = 2)),
+    list(a = 1, b = 2)
+  )
+  expect_equal(
+    merge_lists(list("a" = 1, "b" = 2), list()),
+    list(a = 1, b = 2)
+  )
+  expect_equal(
+    merge_lists(list("a" = 1, "b" = 2), list("b" = 3, "c" = 4)),
+    list(a = 1, b = 2, c = 4)
+  )
+  expect_equal(
+    merge_lists(list("a" = 1, "b" = 2), list("b" = 3, "c" = 4), list("d" = 5)),
+    list(a = 1, b = 2, c = 4, d = 5)
+  )
+})